requirements.txt
idna==3.4
pip==22.0.4
rel==0.4.7
requests==2.28.2
setuptools==58.1.0
tomli==2.0.1
urllib3==1.26.14
websocket-client==1.4.2